Making few changes to make Cocowiwa VS friendly

+++ src/engine/graphics/renderer.cpp
@@ -337,7 +337,7 @@ namespace
 				glViewport(0, 0, dimension.width, dimension.height);
 
 				// these calculations do not need opengl context
-				projection2D = core::maths::Matrix4x4f::ortho(0.f, dimension.width, dimension.height, 0.f, -1.f, 1.f);
+				projection2D = core::maths::Matrix4x4f::ortho(0.f, (float)dimension.width, (float)dimension.height, 0.f, -1.f, 1.f);
 				projection3D = core::maths::Matrix4x4f::perspective(core::maths::make_degree(80.f), float(dimension.width) / float(dimension.height), .125f, 128.f);
 			}
 			//
